requirements.txt
langchain==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.5
chromadb==0.4.22
openai==1.10.0
pypdf==4.0.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
sentence-transformers==2.3.1
